requirements.txt
google-cloud-aiplatform>=1.38.0
click>=8.1.0
PyYAML>=6.0
rich>=13.0.0
google-auth>=2.23.0
google-cloud-storage>=2.10.0
